Split SvgPatternServer
